Mover versoes antigas para legado
