+++ legacy/SEIR_IJ_v8.py
