+++ legacy/rascunho_v1.py
